feat: 최단 경로 서비스 구현

+++ src/main/java/nextstep/subway/path/domain/PathFinder.java
@@ -39,8 +39,8 @@ public class PathFinder {
         return findShortestPath(source, destination).getVertexList();
     }
 
-    public double findShortestPathLength(Station source, Station destination) {
-        return findShortestPath(source, destination).getWeight();
+    public int findShortestPathLength(Station source, Station destination) {
+        return (int) findShortestPath(source, destination).getWeight();
     }
 
     private GraphPath<Station, DefaultWeightedEdge> findShortestPath(Station source, Station destination) {
